+++ lab4/physics.h
@@ -5,19 +5,14 @@
 
 #define STEP_SIZE 1.0 /* the step size use in the integration */
 
-int feuler(pcord_t *a,
-	   float time_step) ;
+int feuler(pcord_t *a, double time_step);
 
-float wall_collide(pcord_t *p,
-		   cord_t wall) ;
+double wall_collide(pcord_t *p, cord_t wall);
 
 
-float collide(pcord_t *p1,
-	      pcord_t *p2) ;
+double collide(pcord_t *p1, pcord_t *p2);
 
-void interact(pcord_t *p1,
-	      pcord_t *p2,
-	      float t) ;
+void interact(pcord_t *p1, pcord_t *p2, double t);
 
 
 #endif
